+++ skops/io/tests/test_external.py
@@ -19,6 +19,7 @@ from sklearn.datasets import make_classification, make_regression
 
 from skops.io import dumps, loads, visualize
 from skops.io.tests._utils import assert_method_outputs_equal, assert_params_equal
+from skops.utils._fixes import make_xgboost_random_forest
 
 # Default settings for generated data
 N_SAMPLES = 30
@@ -235,7 +236,9 @@ class TestXGBoost:
             # This parameter combination is not supported in XGBoost
             return
 
-        estimator = xgboost.XGBRFClassifier(booster=booster, tree_method=tree_method)
+        estimator = make_xgboost_random_forest(
+            xgboost, classifier=True, booster=booster, tree_method=tree_method
+        )
         loaded = loads(dumps(estimator), trusted=trusted)
         assert_params_equal(estimator.get_params(), loaded.get_params())
 
@@ -254,7 +257,9 @@ class TestXGBoost:
             # This parameter combination is not supported in XGBoost
             return
 
-        estimator = xgboost.XGBRFRegressor(booster=booster, tree_method=tree_method)
+        estimator = make_xgboost_random_forest(
+            xgboost, classifier=False, booster=booster, tree_method=tree_method
+        )
         loaded = loads(dumps(estimator), trusted=trusted)
         assert_params_equal(estimator.get_params(), loaded.get_params())